Implement LogoGrid Zeplin specs on live website (#190)
* change Companies component to have 3 columns on small tablet
and fix spacing between rows of logos

* move LogoGrid into Common folder & make naming more generic for use across multiple pages

* make 'Other clients we helped' section dependent on whether logos are entered into Contentful

* use LogoGrid within Technology Partners section of About Us page

* remove ternary which allows LogoGrid to hide itself & default value which allows LogoGrid to display just a div

* rename prop being passed into LogoGrid for greater clarity
